feat: JspView 구현
